♻️ refactor: SponsorPagination과 SponsorSlider의 중복 코드 해결
